### 2022-02-14 #4162 @whedon
[PRE REVIEW]: JOAN: a framework for human-automated vehicle interaction experiments in a virtual reality driving simulator
**Submitting author:** @OlgerSiebinga (<a href="http://orcid.org/0000-0002-5614-1262">Olger Siebinga</a>)
**Repository:** <a href="https://github.com/tud-hri/joan" target ="_blank">https://github.com/tud-hri/joan</a>
**Version:** V1.1.0
