### 2022-02-14 #4162 @danielskatz
👋 @OlgerSiebinga - Thanks for your submission. While I look for an editor, you could work on the possibly missing DOIs that whedon suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@whedon check references` to check again, and the command `@whedon generate pdf` when the references are right to make a new PDF. Whedon commands need to be the first entry in a new comment.
