### 2022-02-15 #4162 @OlgerSiebinga
We want to give a fair warning to all potential reviewers.
Our software uses the open-source autonomous vehicle simulator [CARLA](https://carla.org/), which is needed to properly test our work. Building and installing CARLA will take quite some time (4 hours according to their docs) and you'll need some reasonable hardware (a good graphics card is recommended).
