### 2022-02-15 a1b9d2d
added missing DOIs
