+++ paper/paper.bib
@@ -41,13 +41,13 @@
  lastaccessed={2022-01-28}
  }
 
- @inproceedings{airsim2017fsr,
-  author = {Shital Shah and Debadeepta Dey and Chris Lovett and Ashish Kapoor},
-  title = {AirSim: High-Fidelity Visual and Physical Simulation for Autonomous Vehicles},
-  year = {2017},
-  booktitle = {Field and Service Robotics},
-  eprint = {arXiv:1705.05065},
-  url = {https://arxiv.org/abs/1705.05065}
+@incollection{airsim2017fsr,
+author = {Shah, Shital and Dey, Debadeepta and Lovett, Chris and Kapoor, Ashish},
+doi = {10.1007/978-3-319-67361-5_40},
+pages = {621--635},
+title = {{AirSim: High-Fidelity Visual and Physical Simulation for Autonomous Vehicles}},
+url = {http://link.springer.com/10.1007/978-3-319-67361-5_40},
+year = {2018}
 }
 
  @online{lgsvlsim,
@@ -208,26 +208,24 @@
   language = {en}
 }
 
-
-
-
 @article{di2020_survey,
-  title = {A {{Survey}} on {{Autonomous Vehicle Control}} in the {{Era}} of {{Mixed}}-{{Autonomy}}: {{From Physics}}-{{Based}} to {{AI}}-{{Guided Driving Policy Learning}}},
-  shorttitle = {A {{Survey}} on {{Autonomous Vehicle Control}} in the {{Era}} of {{Mixed}}-{{Autonomy}}},
-  author = {Di, Xuan and Shi, Rongye},
-  year = {2020},
-  month = jul,
-  url = {http://arxiv.org/abs/2007.05156},
-  urldate = {2020-09-14},
-  abstract = {This paper serves as an introduction and overview of the potentially useful models and methodologies from artificial intelligence (AI) into the field of transportation engineering for autonomous vehicle (AV) control in the era of mixed autonomy. It is the first-of-its-kind survey paper to comprehensively review literature in both transportation engineering and AI for mixed traffic modeling. We will discuss state-ofthe-art applications of AI-guided methods, identify opportunities and obstacles, raise open questions, and help suggest the building blocks and areas where AI could play a role in mixed autonomy. We divide the stage of autonomous vehicle (AV) deployment into four phases: the pure HVs, the HV-dominated, the AV-dominated, and the pure AVs. This paper is primarily focused on the latter three phases. Models used for each phase are summarized, encompassing game theory, deep (reinforcement) learning, and imitation learning. While reviewing the methodologies, we primarily focus on the following research questions: (1) What scalable driving policies are to control a large number of AVs in mixed traffic comprised of human drivers and uncontrollable AVs? (2) How do we estimate human driver behaviors? (3) How should the driving behavior of uncontrollable AVs be modeled in the environment? (4) How are the interactions between human drivers and autonomous vehicles characterized? Hopefully this paper will not only inspire our transportation community to rethink the conventional models that are developed in the data-shortage era, but also reach out to other disciplines, in particular robotics and machine learning, to join forces towards creating a safe and efficient mixed traffic ecosystem.},
-  archivePrefix = {arXiv},
-  eprint = {2007.05156},
-  eprinttype = {arxiv},
-  file = {C\:\\Users\\nwmbeckers\\Dropbox\\library-tud\\Di2020_A Survey on Autonomous Vehicle Control in the Era of Mixed-Autonomy.pdf},
-  journal = {arXiv:2007.05156 [cs]},
-  keywords = {⛔ No DOI found,Computer Science - Artificial Intelligence,Computer Science - Robotics},
-  language = {en},
-  primaryClass = {cs}
[…]
