### 2022-02-28 #4162 @humanfactors
Happy to review this. I'm currently on sick leave, so I wouldn't be able to _start_ this until at least next week. As it appears a more intensive workload in terms of setting up, I can't give a determinate date I can complete the review at this stage (though I obviously would get to it as soon as possible).

I also don't have VR, not sure if that is required for tests?
