### 2022-02-28 #4162 @taless474
@humanfactors thank you. You can start in two weeks.
@OlgerSiebinga I think the required hardware is only a good GPU setup not a VR. Is it right?
