### 2022-02-28 #4162 @OlgerSiebinga
That's right! Our work does support experiments in VR, but everything can be used and tested without VR.
