### 2022-03-15 #4162 @humanfactors
I begun my review this week and have spent quite a bit of time trying to validate the functionality.

Unfortunately, I am unable to get CARLA compiled on my computer. I followed the steps specified, but I am having some errors reported on their Github by other users too (related to  `-[install_rpclib]:  - Make sure "CMake" is installed. ` during PythonAPI Make despite CMake being on path). I admit, I do have other versions of Visual Studio Installed (though I am using 2017), but unfortunately I am not able to uninstall these on my work computer. I really wanted to get this to work, but I also do need to balance the requisite debugging for the purpose of only a single test.

I have had no problems installing JOAN itself, but I will be unable to verify the claimed functionality in terms of it interfacing with a CARLA session. @taless474 I do have sufficient information to carry out a review on other criteria, and am happy to complete that. However, I would think we do need an additional reviewer to actually replicate the interface with Carla.

-------------

I should also flag this is a bit unique relative to the other work I have reviewed in JOSS. JOSS reviewer guidelines indicates generally the installation process should be as automated as possible (e.g., automated scripts) with an unambiguous set of tests for validating the functionality of the software. The fact that JOAN depends on the compilation of a secondary piece of software does make this criteria a little vague to interpret. To be clear, this issue is with Carla, not JOAN.
