### 2022-03-17 #4250 @editorialbot
[REVIEW]: JOAN: a framework for human-automated vehicle interaction experiments in a virtual reality driving simulator
**Submitting author:** @OlgerSiebinga (<a href="http://orcid.org/0000-0002-5614-1262">Olger Siebinga</a>)
**Repository:** https://github.com/tud-hri/joan
**Branch with paper.md** (empty if default branch):
**Version:** V1.1.0
