### 2022-03-21 #4250 @humanfactors
Hi @OlgerSiebinga, I've made an Issue regarding my feedback on the JOAN side (rather than paper itself): https://github.com/tud-hri/joan/issues/10

Overall, I think this is an impressive contribution to human factors research! I admit though, I couldn't get Carla installed, and I am not familiar enough with JOSS to know whether the manual compilation and installation is an issue at all.

The only other question I'll raise casually is I noticed there were a few "stale" issues (e.g., https://github.com/tud-hri/joan/issues/3), and might suggest it could be good to have a template to request a reproducible example to replicate the bug etc, or identify a pathway to close these kind of things up. This isn't really specific to the review though, but I was curious as it related to contributing guidelines.

Cheers!
