### 2022-03-23 99f5481
Update introduction.md, addressing JOSS reviewer comments 1.1, 1.2, 4.1
