+++ docs/introduction.md
@@ -1,11 +1,10 @@
 # Introduction
 
-JOAN is a human-automated vehicle simulator specifically built to enable human-in-the-loop experiments. We built JOAN on top of CARLA, a simulator developed for autonomous driving
-research. JOAN's main goal is to provide researchers with an customizable framework and easy to use interface for setting up experiments that involve driving simulation.
+To design automated vehicles (AVs) that can be safely introduced in our mixed traffic, research into human-AV interaction is needed. Driving simulators are invaluable tools, however, existing driving simulators are expensive, implementing new functionalities (e.g., AV control algorithms) can be difficult, and setting up new experiments can be tedious. To address these issues, we present JOAN, a framework for human-AV interaction experiments. JOAN connects to the open-source AV-simulation program Carla and enables quick and easy implementation of human-in-the-loop driving experiments. It supports experiments in VR, a variety of human input devices, and haptic feedback. JOAN is easy to use, flexible, and enables researchers to design, store, and perform human-factors experiments without writing a single line of code.
 
-CARLA provides us with a realistic virtual driving simulator, in which the user can create complex and realistic traffic simulations. CARLA takes care of the physics computations,
-updates of the world and the actors in the world and much more. JOAN plug into this by providing you with means to control an actor through a steering wheel and pedals, create
-experiments, log data, code and test haptic feedback on the steering wheel, etc.
+JOAN builts upon Carla, a open-source game-based driving simulation created to develop AV algorithms. Carla provides a flexible and well-documented API, pre-trained automated driving models, common roadmap standards (OpenDrive), and is actively maintained. While Carla provides possibilities for implementing, training, and evaluating fully automated driving algorithm, it does not explicitly support human-in-the-loop experiments, or make it easy to set up or create such experiments. Carla requires the user to create their own code to use human interfaces, which can be challenging and time-consuming depending on the experimenter's coding abilities.
+
+To address these issues, we created JOAN, an open-source framework for conducting human-in-the-loop driving experiments. JOAN interfaces with Carla, is written in Python, and is fully customizable through code and graphical user interfaces (GUI). JOAN can be used with a variety of human input devices, including game console controllers (e.g.,Xbox or PlayStation), generic USB controllers (e.g., Logitech G920), or a SensoDrive high-fidelity steering wheel (for including haptic feedback). JOAN includes a framework for experiment design to create and execute experiments, and provides reliable data acquisition. These features can be accessed through a user-friendly interface, which means no extensive (knowledge of) programming nor a lot of time is required to set up new experiments.
 
 JOAN mainly consists of modules, which communicate with each other through a News channel. The main modules are:
 
