### 2022-03-23 tud-hri/joan#10 @niekbeckers
Hi @humanfactors,

Thanks for your comprehensive review, we very much appreciate it! We are working on your comments and will reply here on a rolling basis :-)

We've updated the statement of need in readme.md  and the documentation introduction (comments 1.1, 1.2) , better indicating what 'gaps' in Carla JOAN addresses.
