### 2022-04-04 ad54245
improved clarity in installation instructions moved JOAN assets to 4TU data repo provide a list of dependencies up front
