+++ docs/introduction.md
@@ -2,22 +2,38 @@
 
 To design automated vehicles (AVs) that can be safely introduced in our mixed traffic, research into human-AV interaction is needed. Driving simulators are invaluable tools, however, existing driving simulators are expensive, implementing new functionalities (e.g., AV control algorithms) can be difficult, and setting up new experiments can be tedious. To address these issues, we present JOAN, a framework for human-AV interaction experiments. JOAN connects to the open-source AV-simulation program Carla and enables quick and easy implementation of human-in-the-loop driving experiments. It supports experiments in VR, a variety of human input devices, and haptic feedback. JOAN is easy to use, flexible, and enables researchers to design, store, and perform human-factors experiments without writing a single line of code.
 
-JOAN builts upon Carla, a open-source game-based driving simulation created to develop AV algorithms. Carla provides a flexible and well-documented API, pre-trained automated driving models, common roadmap standards (OpenDrive), and is actively maintained. While Carla provides possibilities for implementing, training, and evaluating fully automated driving algorithm, it does not explicitly support human-in-the-loop experiments, or make it easy to set up or create such experiments. Carla requires the user to create their own code to use human interfaces, which can be challenging and time-consuming depending on the experimenter's coding abilities.
+JOAN builds upon Carla, an open-source game-based driving simulation created to develop AV algorithms. Carla provides a flexible and well-documented API, pre-trained automated driving models, common roadmap standards (OpenDrive), and is actively maintained. While Carla provides possibilities for implementing, training, and evaluating fully automated driving algorithm, it does not explicitly support human-in-the-loop experiments, or make it easy to set up or create such experiments. Carla requires the user to create their own code to use human interfaces, which can be challenging and time-consuming depending on the experimenter's coding abilities.
+
+Besides the lack of standardized methods to interface with human input devices, Carla was not designed for repeatable human-in-the-loop experiments. There is no possibility to automatically reset specific scenarios. Nor is there a way to create an experiment with multiple (slightly distinct) conditions. There is also no central data recorder that can easily store data from both the Carla side and from human input devices. 
 
 To address these issues, we created JOAN, an open-source framework for conducting human-in-the-loop driving experiments. JOAN interfaces with Carla, is written in Python, and is fully customizable through code and graphical user interfaces (GUI). JOAN can be used with a variety of human input devices, including game console controllers (e.g.,Xbox or PlayStation), generic USB controllers (e.g., Logitech G920), or a SensoDrive high-fidelity steering wheel (for including haptic feedback). JOAN includes a framework for experiment design to create and execute experiments, and provides reliable data acquisition. These features can be accessed through a user-friendly interface, which means no extensive (knowledge of) programming nor a lot of time is required to set up new experiments.
 
-JOAN mainly consists of modules, which communicate with each other through a News channel. The main modules are:
+JOAN is composed of modules, which communicate with each other through a News channel. All modules have a Settings class which can alter the experiment. The main difference between News and Settings is that News is constantly updated during an experiment Trial while Settings remain constant over a trial. All modules run in their own Python process, meaning that they run completely parallel. The main modules in JOAN are:
 
-- CARLA interface
-- Hardware Manager
-- Data recorder
-- Data Plotter
-- Experiment Manager
-- Haptic Controller Manager
+- __CARLA interface__
+The Carla interface module communicates with Carla. Vehicles (both human driver and NPC) can be spawned from here. It is also the module that connect certain inputs (human inputs or NPC controllers) to specific vehicles 
+
+- __Hardware Manager__
+The hardware manager is used to connect human input devices to JOAN. This is where raw input (e.g., a usb joystick axis value) is translated to a vehicle command (e.g., brake level).
+
+- __Data recorder__
[…]
