### 2022-04-08 8ea41ec
updated contributing guidelines in docs to match CONTRIBUTING.md
