### 2022-04-08 c82250f
Delete .gitlab/issue_templates directory
