+++ .gitlab/issue_templates/feature-request.md
@@ -1,26 +1,0 @@
-<!-- The first three sections: "Problem to solve", "Intended users" and "Proposal", are strongly recommended. Keep in mind that providing complete and relevant information early helps our product team validate the problem and start working on a solution. -->
-
-### Problem to solve
-
-<!-- What problem do_while_running we solve? Try to define the who/what/why of the opportunity as a user story. For example, "As a (who), I want (what), so I can (why/value)." -->
-
-
-### Further details
-
-<!-- Include use cases, benefits, goals, or any other details that will help us understand the problem better. -->
-
-### Proposal
-
-<!-- How are we going to solve the problem? Try to include the user journey! https://about.gitlab.com/handbook/journeys/#user-journey -->
-
-### Testing
-
-<!-- This section needs to be retained and filled in during the workflow planning breakdown phase of this feature proposal, if not earlier.
-
-What risks does this change pose to our availability? How might it affect the quality of the product? What additional test coverage or changes to tests will be needed? 
--->
-<!-- Define both the success metrics and acceptance criteria. Note that success metrics indicate the desired business outcomes, while acceptance criteria indicate when the solution is working correctly. If there is no way to measure success, link to an issue that will implement a way to measure this. -->
-
-### Links / references
-
-/label ~feature
