### 2022-04-12 19e7154
spellchecked all documentation files
