+++ docs/modules-hardwaremanager.md
@@ -1,5 +1,5 @@
 # Module: Hardware Manager
-The Hardware Manager allows hardware inputs to be connected to JOAN. In this module, the connection with the hardware is established and the raw input data is converted to vehicle commands. For example, this is the place where an 8 bit joystick axis value is converted to a steering angle. Hardware inputs are connected to specific vehicles in the Carla interface module. We provide three base classes for hardware inputs: Keyboard, USB Joystick, and SensoDrive. These will be explained on this page. If you want to connect other hardware to JOAN, you will need to modify an existing, or completely write you own, input class.
+The Hardware Manager allows hardware inputs to be connected to JOAN. In this module, the connection with the hardware is established and the raw input data is converted to vehicle commands. For example, this is the place where an 8-bit joystick axis value is converted to a steering angle. Hardware inputs are connected to specific vehicles in the Carla interface module. We provide three base classes for hardware inputs: Keyboard, USB Joystick, and SensoDrive. These will be explained on this page. If you want to connect other hardware to JOAN, you will need to modify an existing, or completely write your own, input class.
  
 In the image below the HardwareManager module is shown how it should look like if you open it up in the 
 `STOPPED` state. (without default settings)
@@ -19,36 +19,36 @@ Whenever adding a keyboard it will automatically open up the keyboard settings d
 
 ![keyboard settings](imgs/modules-hardwaremanager-keyboard-settings.PNG)
 
-The settings in the dialog are quite self explanatory, besides maybe the `sensitivities` section. The higher a sensitivity the faster it will react to you pushing the appropriate button. You can set different keys by pressing the `Set Different Keys` button, the button you can change will then light up green, see the GIF below:
+The settings in the dialog are quite self-explanatory, besides maybe the `sensitivities` section. The higher the sensitivity the faster it will react to you pushing the appropriate button. You can set different keys by pressing the `Set Different Keys` button, the button you can change will then light up green, see the GIF below:
 
 ![Keyboard settings gif](gifs/modules-hardwaremanager-keyboard-settings.gif)
 
 !!! Note
-    If you try to select the same button twice for a different function you will get notified of doing so. You will be able continue, however it can lead to erratic behaviour.
+    If you try to select the same button twice for a different function you will get notified of doing so. You will be able to continue, however, it can lead to erratic behavior.
 
 ### Joysticks
 The second input you can choose from by default is the `Joystick` input. 
 
 !!! Note 
-    The nomenclature `Joystick` may be a bit confusing  since it does not necessarily need to be an actual joystick as you would expect from for example flight simulator. A `Joystick` in JOAN constitutes nothing more than a HID (Human Interface Device). For example, steering wheels by thrustmaster or logitech are also 'Joysticks'.
+    The nomenclature `Joystick` may be a bit confusing since it does not necessarily need to be an actual joystick as you would expect from for example flight simulator. A `Joystick` in JOAN constitutes nothing more than an HID (Human Interface Device). For example, steering wheels by Thrustmaster or Logitech are also 'Joysticks'.
 
 Whenever adding a Joystick device it will automatically open up the joystick settings dialog (if you do it from the `add hardware input` button):
 
 ![Joystick Settings](imgs/modules-hardwaremanager-joystick-settings.PNG)
 
-In here the first thing you do is find your plugged in USB device in the list of `available HID devices`. In this example 
-we have a Xbox Controller connected and will try and preview the inputs. Also notice that there are presets available for
-both Xbox and PlayStation controllers (if you want to use a playstation controller on Windows, install [DS4Windows](http://ds4windows.com/).)
+Here, the first thing you do is find your plugged-in USB device in the list of `available HID devices`. In this example 
[…]
