+++ docs/modules-carlainterface.md
@@ -1,37 +1,37 @@
 # Module: CARLA Interface
 
-The CARLA interface module is one of the core modules, it establishes the connection between JOAN and CARLA. This module is where you can spawn new vehicles (either human driven of NPC) and connect them to control methods (human inputs or NPC controllers). This is also the module that retrieves data from Carla. This data is shared on the modules News channel and can therefore be used in other modules (e.g., current velocity in a controller module), or it can be stored to a csv file in the data recorder. Finally, this module allows you to create _scenarios_. A scenario is an event triggered by a condition in the Carla environment being met. For example, you can use a scenario that starts the movement of an NPC vehicle if the ego vehicle passes a specific point in the scene. 
+The CARLA interface module is one of the core modules, it establishes the connection between JOAN and CARLA. This module is where you can spawn new vehicles (either human-driven or NPC) and connect them to control methods (human inputs or NPC controllers). This is also the module that retrieves data from Carla. This data is shared on the modules News channel and can therefore be used in other modules (e.g., current velocity in a controller module), or it can be stored in a CSV file in the data recorder. Finally, this module allows you to create _scenarios_. A scenario is an event triggered by a condition in the Carla environment being met. For example, you can use a scenario that starts the movement of an NPC vehicle if the ego vehicle passes a specific point in the scene. 
 
 An example screenshot of the module dialog layout in the `READY` state is shown below:
 
 ![carlainterface_dialog](imgs/modules-carlainterface-dialog_ready.PNG)
 
-In this section we'll cover 3 main things namely: `using the module`, `creating your own agent to add` and `creating a scenario`.
+In this section, we'll cover 3 main things namely: `using the module`, `creating your own agent to add`, and `creating a scenario`.
 
 ## <a name="using_carlainterface"></a>Using the Module
 
 ### Carla Connection
 
-Whenever you have included CarlaInterface in your modules, JOAN will try to connect to CARLA at start up. If this process fails you will see the following
+Whenever you have included CarlaInterface in your modules, JOAN will try to connect to CARLA at startup. If this process fails you will see the following
 dialog:
 
 ![Carla interface connection exception](imgs/modules-carlainterface-connection-exception.PNG)
 
-This could happen if you did not start up CARLA in Unreal yet or there are some other connection issues. If connection failed at startup this is not a problem,
+This could happen if you did not start up CARLA in Unreal yet or there are some other connection issues. If the connection failed at startup this is not a problem,
 via the main JOAN window you can still connect to CARLA by pressing the connect button, shown below:
 
 ![Carla Interface Tab](imgs/modules-carlainterface-tab-in-main.PNG)
 
 If JOAN is already connected the connect button will not be available but the disconnect button is enabled. One last thing remains to be said about the CARLA
-connection. Due to the fact that we use multiprocessing we need to connect to CARLA again when the module process is created (this is done after
+connection. Because we use multiprocessing, we need to connect to CARLA again when the module process is created (this is done after
 clicking `get ready`). This makes sure we have access to the CARLA pythonapi function within this process. There is however a big note to this:
 
 !!! Important 
-    Whenever you are already in the `READY` state, and you stop CARLA from the unreal side, this will lead to erratic behaviour in JOAN. So never stop your experiments via the quit button in Unreal. Ideally you use the 'play' button in unreal once before starting up JOAN and stop it again after you quit JOAN.
[…]

[…]
