### 2022-04-12 tud-hri/joan#10 @OlgerSiebinga
Hi @humanfactors,
As @niekbeckers already said, thanks for the many useful and constructive comments! By now, we have addressed most of them. I tried to address and explain the three last open checkboxes below. Let me know if you agree or have any other questions.
### The Build Context
We agree that building Carla is a tedious process with a lot of warnings and conditions. The problem here is that we also don’t fully know to what extent these specific versions of dependencies really matter. The instructions to build Carla are heavily based on their own documentation, and instead of trying to also develop extra tools for building Carla we prefer to link users to their documentation.
### Example of usage
I have updated the [introduction page](https://joan.readthedocs.io/en/latest/introduction/) in the documentation to give more details on how we think JOAN can be best used. I’ve also added a link to a research paper that actually used JOAN. Is that what you had in mind?
### Tests
The testing is a difficult part because we are trying to validate if a human-in-the-loop experiment is working as intended. So comparing (human input) values to a pre-recorded CSV is not going to be of any help. What we are actually testing here boils down to three things:

-	Does the connection with Carla work?
-	Does the hardware input work?
-	Does JOAN’s underlying structure (with settings, news, and a bunch of separate Python processes) work?

If any of these things fail, the user will get an exception or clearly unexpected behavior (e.g., if the human input steering angles are larger than expected). So, we think comparing values is not really necessary to test if JOAN works as expected.

Thanks again for the extensive review of these topics!
