### 2022-04-12 #4250 @OlgerSiebinga
Hi All,

We have been working on the feedback in https://github.com/tud-hri/joan/issues/10, and we have addressed most of it. Hopefully, we can close that issue soon.  @humanfactors, thanks for summarizing your feedback in that issue, it was really helpful.

@taless474, I saw that there is only a checklist for @humanfactors in this issue. Will @bnriiitb get a checklist once @humanfactors is done? Or was the editorialbot confused when setting up this review issue?
