### 2022-04-28 tud-hri/joan#10 @humanfactors
Few quick follow ups:

- Could you ensure all Dropbox links coverted to the AWS (some here https://joan.readthedocs.io/en/latest/example-usage-and-testing/)
- The quote on the home page runs over https://joan.readthedocs.io/en/latest/

Otherwise changes look good!
