### 2022-05-16 #4250 @humanfactors
Hi apologies for the delays, I've had COVID and been quite unwell.

A few minor tasks on the software paper:

- References are appearing incorrect, the authors do need to check up on the reference syntax (e.g., double brackets L41, L97)
- SensoDrive ref missing date

As indicated prior, I was unable to install Carla side of the software, so I am unable to actually check off the other tasks. I was waiting for the other review to report how they got along.

Cheers!
