### 2022-05-16 ecb2c57
minor update to references in paper
