+++ paper/paper.bib
@@ -31,14 +31,16 @@
  title={Driving Simulation and SILAB},
  url={https://wivw.de/en/silab},
  author={WIVW},
- lastaccessed={2020-09-30}
+ note={visited on: 2020-09-30},
+ year={2020}
  }
 
  @online{sensodrive,
  title={SensoDrive Force Feedback},
  url={https://www.sensodrive.de/products/force-feedback-products.php},
  author={SensoDrive},
- lastaccessed={2022-01-28}
+ note={visited on: 2022-01-28},
+ year={2022}
  }
 
 @incollection{airsim2017fsr,
@@ -54,7 +56,8 @@ year = {2018}
  title={LGSVL Simulator},
  url={https://www.lgsvlsimulator.com/},
  author={Advanced Platform Lab at the LG Electronics America R\&D Center},
- lastaccessed={2020-09-30},
+ note={visited on: 2020-09-30},
+ year={2020}
  }
 
 @techreport{kearney2018_multimodal,
@@ -63,11 +66,8 @@ year = {2018}
   year = {2018},
   pages = {31},
   file = {C\:\\Users\\nwmbeckers\\Zotero\\storage\\IWWH65VN\\Kearney and Noyce - Multi-Modal Distributed Simulation Combining Cars,.pdf},
-  keywords = {⛔ No DOI found},
   language = {en}
 }
-
-
 
 @inproceedings{dosovitskiy2017_carla,
   title = {{{CARLA}}: {{An Open Urban Driving Simulator}}},
@@ -79,7 +79,6 @@ year = {2018}
   url = {http://proceedings.mlr.press/v78/dosovitskiy17a/dosovitskiy17a.pdf},
   abstract = {We introduce CARLA, an open-source simulator for autonomous driving research. CARLA has been developed from the ground up to support development, training, and validation of autonomous urban driving systems. In addition to open-source code and protocols, CARLA provides open digital assets (urban layouts, buildings, vehicles) that were created for this purpose and can be used freely. The simulation platform supports flexible specification of sensor suites and environmental conditions. We use CARLA to study the performance of three approaches to autonomous driving: a classic modular pipeline, an endto-end model trained via imitation learning, and an end-to-end model trained via reinforcement learning. The approaches are evaluated in controlled scenarios of increasing difficulty, and their performance is examined via metrics provided by CARLA, illustrating the platform's utility for autonomous driving research.},
   file = {C\:\\Users\\nwmbeckers\\Dropbox\\library-tud\\Dosovitskiy2017_CARLA.pdf},
-  keywords = {⛔ No DOI found},
   language = {en}
 }
 
@@ -104,7 +103,6 @@ year = {2018}
   title = {State-of-the-Art Driving Simulators, a Literature Survey},
   author = {Slob, JJ},
   year = {2008},
-  keywords = {⛔ No DOI found},
   number = {107},
   type = {{{DCT}}}
 }
