### 2022-05-16 d49511a
minor update to references in paper
