+++ paper/paper.bib
@@ -104,7 +104,7 @@ year = {2018}
   author = {Slob, JJ},
   year = {2008},
   number = {107},
-  type = {{{DCT}}}
+  type = {DCT report 107}
 }
 
 @misc{qtframework,
