### 2022-05-16 a601064
minor update to references in paper
