### 2022-05-31 #4250 @humanfactors
Hi @taless474 and @OlgerSiebinga,

My review is completed insofar as I have checked off every item that I possibly can. I am satisfied with the manuscript and all changes made by the authors. As mentioned previously, I was unable to install the CARLA software and I am not sure what the editorial protocol is in such situations. If another independent review can confirm they can compile the software, then I guess I can also check off those items.

I spent a good day or so attempting the detailed compilation and installation process, and ran into issues with the CARLA compilation process (https://github.com/openjournals/joss-reviews/issues/4162#issuecomment-1067511114). To be totally transparent, I cannot commit further resources to that aspect of the review given the substantial investment that would be required on my part to validate a single test.

Apologies I cannot be of more help, but if @bnriiitb similarly does not have the capacity, perhaps we should be identifying a third reviewer who is able to compile and expedite the review process?
