### 2022-06-07 fd332c8
Included check that opens carla if not opened
