### 2022-06-07 d1b30b0
Included check that opens carla if not opened
