### 2022-06-23 #4250 @OlgerSiebinga
Hi All,

We paused the development of JOAN pending this review, but it has been a while now. So, we would like to resume the development on the `develop` and `demo` branches (we plan to make some substantial changes to the docs and procedures to allow JOAN to run with newer versions of CARLA). Please use the `master` branch for this review, we will not touch it until the review is done. That way, the review can continue on the same version of JOAN while we work on newer versions in parallel. Just wanted to let you know, thanks!
