### 2022-06-29 47fc6e6
[BUGFIX] Fixed bug that prevented time stamps from being saved properly to csv files
