### 2022-07-01 0884fb0
Set up initial version of logitech steering wheel hardware type
