+++ modules/hardwaremanager/hardwaremanager_sharedvariables.py
@@ -15,61 +15,7 @@ class HardwareManagerSharedVariables(ModuleSharedVariables):
         self.inputs = {}
 
 
-class KeyboardSharedVariables(SharedVariables):
-    """"
-    This class contains all the variables that are shared between the seperate hardware communication core and the
-    main JOAN core.
-    """
-
-    def __init__(self):
-        self._steering_angle = mp.Value(c_float, 0.0)
-        self._throttle = mp.Value(c_float, 0.0)
-        self._brake = mp.Value(c_float, -9.9)
-        self._reverse = mp.Value(c_bool, False)
-        self._handbrake = mp.Value(c_bool, False)
-
-    @property
-    def steering_angle(self):
-        return self._steering_angle.value
-
-    @steering_angle.setter
-    def steering_angle(self, val):
-        self._steering_angle.value = val
-
-    @property
-    def throttle(self):
-        return self._throttle.value
-
-    @throttle.setter
-    def throttle(self, val):
-        self._throttle.value = val
-
-    @property
-    def brake(self):
-        return self._brake.value
-
-    @brake.setter
-    def brake(self, val):
-        self._brake.value = val
-
-    @property
-    def reverse(self):
-        return self._reverse.value
-
-    @reverse.setter
-    def reverse(self, val):
-        self._reverse.value = val
-
-    @property
-    def handbrake(self):
-        return self._handbrake.value
-
-    @handbrake.setter
-    def handbrake(self, val):
-        self._handbrake.value = val
-
-
-class JoystickSharedVariables(SharedVariables):
+class GenericHardwareSharedVariables(SharedVariables):
     """"
     This class contains all the variables that are shared between the seperate hardware communication core and the
     main JOAN core.
