### 2022-07-01 0e93799
Set up initial version of logitech steering wheel hardware type
