+++ requirements.txt
@@ -14,3 +14,4 @@ mkdocs
 colour~=0.1.5
 wres~=1.0.3
 pathlib~=1.0.1
+logitech-steering-wheel
