### 2022-07-01 42a71f5
Revert "Set up initial version of logitech steering wheel hardware type"
This reverts commit 0e93799c6ca7a031ff2609884882c383b15fa1f8.
