+++ modules/hapticcontrollermanager/hapticcontrollermanager_controllers/tradedcontroller.py
@@ -1,0 +1,499 @@
+import math
+import os
+
+import numpy as np
+import pandas as pd
+from PyQt5 import QtWidgets, uic
+
+from core.statesenum import State
+from modules.hapticcontrollermanager.hapticcontrollermanager_controllertypes import HapticControllerTypes
+from tools import LowPassFilterBiquad
+from tools.haptic_controller_tools import find_closest_node
+
+
+class TradedControllerSettingsDialog(QtWidgets.QDialog):
+    def __init__(self, settings, module_manager, parent=None):
+        super().__init__(parent)
+
+        self.traded_controller_settings = settings
+        self.module_manager = module_manager
+
+        uic.loadUi(os.path.join(os.path.dirname(os.path.realpath(__file__)), "ui/tradedcontroller_settings_ui.ui"), self)
+        self._path_trajectory_directory = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'trajectories')
+        self.btn_apply_parameters.clicked.connect(self.update_parameters)
+        self.module_manager.state_machine.add_state_change_listener(self.handle_state_change)
+
+        self.update_trajectory_list()
+        self._display_values()
+
+        self.handle_state_change()
+
+    def handle_state_change(self):
+        # disable changing of trajectory while running (would be quite dangerous)
+        if self.module_manager.state_machine.current_state == State.RUNNING:
+            self.cmbbox_hcr_selection.setEnabled(False)
+            self.cmbbox_hcr_selection.blockSignals(True)
+        else:
+            self.cmbbox_hcr_selection.setEnabled(True)
+            self.cmbbox_hcr_selection.blockSignals(False)
+
+    def update_parameters(self):
+        # self.slider_loha.setValue(self.spin_loha.value())
+        self.traded_controller_settings.k_y = float(self.edit_k_y.text())
+        self.traded_controller_settings.k_psi = float(self.edit_k_psi.text())
+        self.traded_controller_settings.kd_y = float(self.edit_kd_y.text())
+        self.traded_controller_settings.kd_psi = float(self.edit_kd_psi.text())
+        self.traded_controller_settings.alpha = float(self.edit_alpha.text())
+        self.traded_controller_settings.tau_th = float(self.edit_tau_th.text())
+        self.traded_controller_settings.trajectory_name = self.cmbbox_hcr_selection.itemText(
+            self.cmbbox_hcr_selection.currentIndex())
+
+        try:
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_y = self.traded_controller_settings.k_y
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_psi = self.traded_controller_settings.k_psi
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_y = self.traded_controller_settings.kd_y
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_psi = self.traded_controller_settings.kd_psi
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].alpha = self.traded_controller_settings.alpha
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].tau_th = self.traded_controller_settings.tau_th
+        except:
+            pass
+
+        self._display_values()
+
+
+    def accept(self):
+        # self.slider_loha.setValue(self.spin_loha.value())
+        self.traded_controller_settings.k_y = float(self.edit_k_y.text())
+        self.traded_controller_settings.k_psi = float(self.edit_k_psi.text())
+        self.traded_controller_settings.kd_y = float(self.edit_kd_y.text())
[…]
