### 2022-07-08 1c5d492
A lot has happened, I made a traded controller now, perhaps need to integrate the torque estimation on steering wheel. Also fixed the steering wheel ratio. Was something like 6.4, more realistic as +- 14. Still need to check if the shared controllers are implemented okayish
