+++ modules/carlainterface/scenarios/scenario_stoptrialatspawnpoint.py
@@ -1,5 +1,6 @@
 from modules.carlainterface.carlainterface_process import CarlaInterfaceProcess
-from tools.carlaimporter import carla
+# from tools.carlaimporter import carla
+import carla
 
 class ScenarioStopTrialAtLocation:
 
