### 2022-07-13 87be000
changed some controller settings
