+++ modules/hapticcontrollermanager/hapticcontrollermanager_controllers/fdcacontroller.py
@@ -155,7 +155,7 @@ class FDCAControllerProcess:
         self.carla_interface_settings = carla_interface_settings
         self.shared_variables = shared_variables
         self._trajectory = None
-        self.t_lookahead = 0
+        self.t_lookahead = 0.8
         self._path_trajectory_directory = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'trajectories')
         self.load_trajectory()
 
@@ -200,7 +200,7 @@ class FDCAControllerProcess:
         :param vel_car:
         :return:
         """
-        pos_car = pos_car + vel_car * self.t_lookahead  # linear extrapolation, should be updated
+        # pos_car = pos_car + vel_car * self.t_lookahead  # linear extrapolation, should be updated
 
         # Find waypoint index of the point that the car would be in the future (compared to own driven trajectory)
         index_closest_waypoint = find_closest_node(pos_car, self._trajectory[:, 1:3])
@@ -263,7 +263,7 @@ class FDCAControllerProcess:
 
     # TODO: I don't buy that this actually works!
     def _get_reference_sw_angle(self, t_ahead, location, velocity):
-        future_location = location + velocity * t_ahead
+        future_location = location #+ velocity * t_ahead
 
         idx = find_closest_node(future_location, self._trajectory[:, 1:3])
         if idx >= len(self._trajectory) - 20:
@@ -273,7 +273,7 @@ class FDCAControllerProcess:
 
         # the trajectory is recorded in unitless steering angles (verify this in the csv)
         # so, we need to convert this to radians. First, multiply with 450 (1, -1) [-] = (450,-450) [deg]
-        sw_angle_ff_des = math.radians(self._trajectory[idx1, 3] * 450)
+        sw_angle_ff_des = math.radians(self._trajectory[idx1, 3] * 70 * 13)
 
         return sw_angle_ff_des
 
@@ -287,105 +287,119 @@ class FDCAControllerProcess:
         """
 
         for agent_settings in carla_interface_settings.agents.values():
-            if agent_settings.selected_controller == self.settings.__str__():
-                if 'SensoDrive' in agent_settings.selected_input:
-                    stiffness = hardware_manager_shared_variables.inputs[agent_settings.selected_input].auto_center_stiffness
+            try:
+                if agent_settings.selected_controller == self.settings.__str__():
+                    if 'SensoDrive' in agent_settings.selected_input:
 
-                    delta_t = time_step_in_ns / 1e9  # [s]
+                        stiffness = hardware_manager_shared_variables.inputs[agent_settings.selected_input].auto_center_stiffness
+                        delta_t = time_step_in_ns / 1e9  # [s]
+                        sw_angle = hardware_manager_shared_variables.inputs[agent_settings.selected_input].steering_angle
 
-                    sw_angle = hardware_manager_shared_variables.inputs[agent_settings.selected_input].steering_angle
+                        # extract data from CARLA
+                        # CARLA coordinate system (left-handed coordinate system)
+                        # X: forward
+                        # Y: right
+                        # Z: upward
+                        # Psi (heading): left-hand z-axis positive (yaw to the right is positive)
+                        # torque: rightward rotation is positive
 
-                    # extract data from CARLA
-                    # CARLA coordinate system (left-handed coordinate system)
-                    # X: forward
-                    # Y: right
-                    # Z: upward
-                    # Psi (heading): left-hand z-axis positive (yaw to the right is positive)
-                    # torque: rightward rotation is positive
+                        pos_car = np.array([carlainterface_shared_variables.agents[agent_settings.__str__()].transform[0],
+                                            carlainterface_shared_variables.agents[agent_settings.__str__()].transform[1]])
[…]
