### 2022-07-13 6b70f03
changed some controller settings
