### 2022-08-02 #4250 @humanfactors
@taless474 It is plausible that @bnriiitb may not have notifications in this thread on, as I don't believe he's made an active response since 18 Mar (> 4.5 months). Is there a protocol for seeking a third reviewer to replicate the process?

Again, I am confident that the actual software that is being published can install/run, just not how it interacts with Carla.
