+++ modules/hapticcontrollermanager/hapticcontrollermanager_controllers/tradedcontroller.py
@@ -7,6 +7,7 @@ from PyQt5 import QtWidgets, uic
 
 from core.statesenum import State
 from modules.hapticcontrollermanager.hapticcontrollermanager_controllertypes import HapticControllerTypes
+from modules.hapticcontrollermanager.hapticcontrollermanager_controllers.fdcacontroller import FDCAController
 from tools import LowPassFilterBiquad
 from tools.haptic_controller_tools import find_closest_node
 
@@ -22,10 +23,8 @@ class TradedControllerSettingsDialog(QtWidgets.QDialog):
         self._path_trajectory_directory = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'trajectories')
         self.btn_apply_parameters.clicked.connect(self.update_parameters)
         self.module_manager.state_machine.add_state_change_listener(self.handle_state_change)
-
         self.update_trajectory_list()
         self._display_values()
-
         self.handle_state_change()
 
     def handle_state_change(self):
@@ -39,22 +38,27 @@ class TradedControllerSettingsDialog(QtWidgets.QDialog):
 
     def update_parameters(self):
         # self.slider_loha.setValue(self.spin_loha.value())
-        self.traded_controller_settings.k_y = float(self.edit_k_y.text())
-        self.traded_controller_settings.k_psi = float(self.edit_k_psi.text())
-        self.traded_controller_settings.kd_y = float(self.edit_kd_y.text())
-        self.traded_controller_settings.kd_psi = float(self.edit_kd_psi.text())
+        # self.traded_controller_settings.k_y = float(self.edit_k_y.text())
+        # self.traded_controller_settings.k_psi = float(self.edit_k_psi.text())
+        # self.traded_controller_settings.kd_y = float(self.edit_kd_y.text())
+        # self.traded_controller_settings.kd_psi = float(self.edit_kd_psi.text())
         self.traded_controller_settings.alpha = float(self.edit_alpha.text())
         self.traded_controller_settings.tau_th = float(self.edit_tau_th.text())
+        if self.checkInvert.isChecked():
+            self.traded_controller_settings.gamma = 1
+        else:
+            self.traded_controller_settings.gamma = -1
         self.traded_controller_settings.trajectory_name = self.cmbbox_hcr_selection.itemText(
             self.cmbbox_hcr_selection.currentIndex())
 
         try:
-            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_y = self.traded_controller_settings.k_y
-            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_psi = self.traded_controller_settings.k_psi
-            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_y = self.traded_controller_settings.kd_y
-            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_psi = self.traded_controller_settings.kd_psi
+            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_y = self.traded_controller_settings.k_y
+            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_psi = self.traded_controller_settings.k_psi
+            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_y = self.traded_controller_settings.kd_y
+            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_psi = self.traded_controller_settings.kd_psi
             self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].alpha = self.traded_controller_settings.alpha
[…]

[…]
