+++ modules/hapticcontrollermanager/hapticcontrollermanager_controllers/tradedcontroller.py
@@ -37,11 +37,6 @@ class TradedControllerSettingsDialog(QtWidgets.QDialog):
             self.cmbbox_hcr_selection.blockSignals(False)
 
     def update_parameters(self):
-        # self.slider_loha.setValue(self.spin_loha.value())
-        # self.traded_controller_settings.k_y = float(self.edit_k_y.text())
-        # self.traded_controller_settings.k_psi = float(self.edit_k_psi.text())
-        # self.traded_controller_settings.kd_y = float(self.edit_kd_y.text())
-        # self.traded_controller_settings.kd_psi = float(self.edit_kd_psi.text())
         self.traded_controller_settings.alpha = float(self.edit_alpha.text())
         self.traded_controller_settings.tau_th = float(self.edit_tau_th.text())
         if self.checkInvert.isChecked():
@@ -52,10 +47,6 @@ class TradedControllerSettingsDialog(QtWidgets.QDialog):
             self.cmbbox_hcr_selection.currentIndex())
 
         try:
-            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_y = self.traded_controller_settings.k_y
-            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_psi = self.traded_controller_settings.k_psi
-            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_y = self.traded_controller_settings.kd_y
-            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_psi = self.traded_controller_settings.kd_psi
             self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].alpha = self.traded_controller_settings.alpha
             self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].tau_th = self.traded_controller_settings.tau_th
             self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].gamma = self.traded_controller_settings.gamma
@@ -66,25 +57,16 @@ class TradedControllerSettingsDialog(QtWidgets.QDialog):
 
 
     def accept(self):
-        # self.slider_loha.setValue(self.spin_loha.value())
-        # self.traded_controller_settings.k_y = float(self.edit_k_y.text())
-        # self.traded_controller_settings.k_psi = float(self.edit_k_psi.text())
-        # self.traded_controller_settings.kd_y = float(self.edit_kd_y.text())
-        # self.traded_controller_settings.kd_psi = float(self.edit_kd_psi.text())
         self.traded_controller_settings.alpha = float(self.edit_alpha.text())
         self.traded_controller_settings.tau_th = float(self.edit_tau_th.text())
         if self.checkInvert.isChecked():
-            self.traded_controller_settings.gamma = 1
-        else:
             self.traded_controller_settings.gamma = -1
+        else:
+            self.traded_controller_settings.gamma = 1
         self.traded_controller_settings.trajectory_name = self.cmbbox_hcr_selection.itemText(
             self.cmbbox_hcr_selection.currentIndex())
 
         try:
-            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_y = self.traded_controller_settings.k_y
-            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].k_psi = self.traded_controller_settings.k_psi
-            # self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].kd_y = self.traded_controller_settings.kd_y
[…]
