### 2022-08-12 2773ac5
Update to Carla 0.9.13 (#13)
* Included check that opens carla if not opened

* A lot has happened, I made a traded controller now, perhaps need to integrate the torque estimation on steering wheel.
Also fixed the steering wheel ratio. Was something like 6.4, more realistic as +- 14.
Still need to check if the shared controllers are implemented okayish

* changed some controller settings

* Fixed both FDCA and TC controller

* minor changes to controllers

* cruise control now also works

* added some comments

* changed username

* Updated documentation

Co-authored-by: loranjenkins <loranbog
[…]
