+++ modules/hapticcontrollermanager/hapticcontrollermanager_controllers/tradedcontroller.py
@@ -1,0 +1,305 @@
+import math
+import os
+
+import numpy as np
+import pandas as pd
+from PyQt5 import QtWidgets, uic
+
+from core.statesenum import State
+from modules.hapticcontrollermanager.hapticcontrollermanager_controllertypes import HapticControllerTypes
+from modules.hapticcontrollermanager.hapticcontrollermanager_controllers.fdcacontroller import FDCAController
+from tools import LowPassFilterBiquad
+from tools.haptic_controller_tools import find_closest_node
+
+
+class TradedControllerSettingsDialog(QtWidgets.QDialog):
+    def __init__(self, settings, module_manager, parent=None):
+        super().__init__(parent)
+
+        self.traded_controller_settings = settings
+        self.module_manager = module_manager
+
+        uic.loadUi(os.path.join(os.path.dirname(os.path.realpath(__file__)), "ui/tradedcontroller_settings_ui.ui"), self)
+        self._path_trajectory_directory = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'trajectories')
+        self.btn_apply_parameters.clicked.connect(self.update_parameters)
+        self.module_manager.state_machine.add_state_change_listener(self.handle_state_change)
+        self.update_trajectory_list()
+        self._display_values()
+        self.handle_state_change()
+
+    def handle_state_change(self):
+        # disable changing of trajectory while running (would be quite dangerous)
+        if self.module_manager.state_machine.current_state == State.RUNNING:
+            self.cmbbox_hcr_selection.setEnabled(False)
+            self.cmbbox_hcr_selection.blockSignals(True)
+        else:
+            self.cmbbox_hcr_selection.setEnabled(True)
+            self.cmbbox_hcr_selection.blockSignals(False)
+
+    def update_parameters(self):
+        self.traded_controller_settings.alpha = float(self.edit_alpha.text())
+        self.traded_controller_settings.tau_th = float(self.edit_tau_th.text())
+        if self.checkInvert.isChecked():
+            self.traded_controller_settings.gamma = -1
+        else:
+            self.traded_controller_settings.gamma = 1
+        self.traded_controller_settings.trajectory_name = self.cmbbox_hcr_selection.itemText(
+            self.cmbbox_hcr_selection.currentIndex())
+
+        try:
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].alpha = self.traded_controller_settings.alpha
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].tau_th = self.traded_controller_settings.tau_th
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].gamma = self.traded_controller_settings.gamma
+        except:
+            pass
+
+        self._display_values()
+
+
+    def accept(self):
+        self.traded_controller_settings.alpha = float(self.edit_alpha.text())
+        self.traded_controller_settings.tau_th = float(self.edit_tau_th.text())
+        if self.checkInvert.isChecked():
+            self.traded_controller_settings.gamma = -1
+        else:
+            self.traded_controller_settings.gamma = 1
+        self.traded_controller_settings.trajectory_name = self.cmbbox_hcr_selection.itemText(
+            self.cmbbox_hcr_selection.currentIndex())
+
+        try:
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].alpha = self.traded_controller_settings.alpha
+            self.module_manager.shared_variables.haptic_controllers[self.traded_controller_settings.identifier].tau_th = self.traded_controller_settings.tau_th
[…]
