+++ modules/hapticcontrollermanager/hapticcontrollermanager_controllers/fdcacontroller.py
@@ -3,7 +3,9 @@ import os
 
 import numpy as np
 import pandas as pd
+import scipy as cp
 from PyQt5 import QtWidgets, uic
+import matplotlib.pyplot as plt
 
 from core.statesenum import State
 from modules.hapticcontrollermanager.hapticcontrollermanager_controllertypes import HapticControllerTypes
@@ -23,15 +25,7 @@ class FDCAControllerSettingsDialog(QtWidgets.QDialog):
 
         self.btnbox_fdca_controller_settings.button(
             self.btnbox_fdca_controller_settings.RestoreDefaults).clicked.connect(self._set_default_values)
-        self.slider_loha.valueChanged.connect(self._update_loha_slider_label)
         self.btn_apply_parameters.clicked.connect(self.update_parameters)
-
-        # hardcode lookahead time if someone needs it
-        self.t_lookahead = 0
-
-        self._loha_resolution = 50
-        self.slider_loha.setMaximum(self._loha_resolution)
-        self.spin_loha.setMaximum(self._loha_resolution)
 
         self.module_manager.state_machine.add_state_change_listener(self.handle_state_change)
 
@@ -50,12 +44,12 @@ class FDCAControllerSettingsDialog(QtWidgets.QDialog):
             self.cmbbox_hcr_selection.blockSignals(False)
 
     def update_parameters(self):
-        self.slider_loha.setValue(self.spin_loha.value())
+        # self.slider_loha.setValue(self.spin_loha.value())
+        self.fdca_controller_settings.loha = float(self.edit_loha.text())
         self.fdca_controller_settings.k_y = float(self.edit_k_y.text())
         self.fdca_controller_settings.k_psi = float(self.edit_k_psi.text())
         self.fdca_controller_settings.lohs = float(self.edit_lohs.text())
         self.fdca_controller_settings.sohf = float(self.edit_sohf.text())
-        self.fdca_controller_settings.loha = float(self.slider_loha.value())
         self.fdca_controller_settings.trajectory_name = self.cmbbox_hcr_selection.itemText(
             self.cmbbox_hcr_selection.currentIndex())
 
@@ -70,24 +64,14 @@ class FDCAControllerSettingsDialog(QtWidgets.QDialog):
 
         self._display_values()
 
-    def _update_loha_slider_label(self):
-        self.spin_loha.setValue(self.slider_loha.value())
-        if self.checkbox_tuning_loha.isChecked():
-            self.fdca_controller_settings.loha = float(self.slider_loha.value())
-            self.lbl_loha.setText(str(self.fdca_controller_settings.loha))
-            self.lbl_loha_deg.setText(str(round(math.radians(self.fdca_controller_settings.loha), 3)))
-            try:
-                self.module_manager.shared_variables.haptic_controllers[self.fdca_controller_settings.identifier].loha = self.fdca_controller_settings.loha
-            except:
-                pass
-
     def accept(self):
-        self.slider_loha.setValue(self.spin_loha.value())
+        # self.slider_loha.setValue(self.spin_loha.value())
+        self.fdca_controller_settings.loha = float(self.edit_loha.text())
         self.fdca_controller_settings.k_y = float(self.edit_k_y.text())
         self.fdca_controller_settings.k_psi = float(self.edit_k_psi.text())
         self.fdca_controller_settings.lohs = float(self.edit_lohs.text())
         self.fdca_controller_settings.sohf = float(self.edit_sohf.text())
-        self.fdca_controller_settings.loha = float(self.slider_loha.value())
+        self.fdca_controller_settings.loha = float(self.edit_loha.text())
[…]
