+++ modules/carlainterface/carlainterface_agentclasses/npc_vehicle.py
@@ -3,7 +3,8 @@ import numpy as np
 from enum import Enum
 import copy
 
-from tools.carlaimporter import carla
+# from tools.carlaimporter import carla
+import carla
 
 from PyQt5 import uic, QtWidgets
 from modules.carlainterface.carlainterface_agenttypes import AgentTypes
@@ -162,7 +163,7 @@ class NPCVehicleProcess:
         self.set_shared_variables()
 
     def destroy(self):
-        if hasattr(self, 'spawned_vehicle'):
+        if hasattr(self, 'spawned_vehicle') and self.spawned_vehicle.is_alive:
             self.spawned_vehicle.destroy()
 
     def set_shared_variables(self):

[…]
