+++ modules/carlainterface/carlainterface_manager.py
@@ -9,8 +9,7 @@ from core.statesenum import State
 from modules.carlainterface.carlainterface_agenttypes import AgentTypes
 from modules.joanmodules import JOANModules
 
-# from tools.carlaimporter import carla
-import carla
+from tools.carlaimporter import carla
 
 
 class CarlaInterfaceManager(ModuleManager):
