+++ modules/carlainterface/carlainterface_process.py
@@ -11,8 +11,8 @@ from modules.joanmodules import JOANModules
 if platform.system() == 'Windows':
     import wres
 
-# from tools.carlaimporter import carla
-import carla
+from tools.carlaimporter import carla
+
 
 def connect_carla(host='localhost', port=2000):
     """
@@ -88,9 +88,6 @@ class CarlaInterfaceProcess(ModuleProcess):
         for key, value in self._settings_as_object.agents.items():
             self.agent_objects[key] = AgentTypes(value.agent_type).process(self, settings=value, shared_variables=self._module_shared_variables.agents[key])
             time.sleep(0.1)  # short sleep, such that the camera in CARLA is attached properly to the first spawned car
-
-    def start(self):
-        super().start()
 
     def run(self):
         """
