+++ docs/setup-carla-windows.md
@@ -1,35 +1,234 @@
 # Setting up CARLA for Windows
 
 !!! Note 
-    Make sure to use [the windows build version CARLA 0.9.13](https://carla.readthedocs.io/en/0.9.13/build_windows){target="_blank"}. 
-    
-First, to install JOAN on Windows, Carla must be installed. This installation procedure is fully explained in the documentation [to build CARLA 0.9.13 for windows](https://carla.readthedocs.io/en/0.9.13/build_windows/){target="_blank"} (the last version we tested JOAN with).
-Below, some additional notes on the installation of Carla 0.9.13 are provided to help in the installation process. 
-Make sure you have completed all the steps in the installation before moving on to the installation of JOAN!
+    If you are a TU Delft student working on one of the TUD computers with CARLA installed, you can skip these instructions. Go to the [setup guide for installing on TUD shared hardware](setup-on-tud-shared-hardware.md).
+
+This guide explains how to build CARLA on Windows, a pre-packaged version of CARLA is also available from their website but that version does not allow advanced customization and development. These instructions are based on the awesome documentation provided by the CARLA team. We tried to describe the steps a bit for our specific use. However, feel free to follow the CARLA installation notes (select the right version). If you get stuck somewhere, or you want to learn more, you could have a look at [their documentation](https://carla.readthedocs.io/en/latest/){target="_blank"} (make sure to select the correct version that corresponds to the CARLA version you want to use).
+
 !!! Note 
-    It may also be a good idea to read some of the additional documentation of [CARLA 0.9.13](https://carla.readthedocs.io/en/0.9.13/){target="_blank"}. 
-
-
+    For the most up-to-date installation instructions for CARLA, check out [their documentation](https://carla.readthedocs.io/en/latest/){target="_blank"}. The instructions
+below are from [CARLA 0.9.9](https://carla.readthedocs.io/en/0.9.9/){target="_blank"} (the last version we tested JOAN with).
 
 ---
 
-## Additional Notes
+## Introduction
 
-First of all, this installation is a game of patience. Assume that installing everything may take 1 full workday so make sure to provide yourself with some snacks and make sure you sit comfortable since this will take a while!
+## Hardware Requirements
 
+* __x64 system__: The simulator should run in any 64 bits Windows system.
+* __50GB disk space__: Installing all the software needed (Unreal, VS 2017, CARLA, and more) will require quite a lot of space. Make sure to have about 50GB of space (especially
+  since you will also need a Visual Studio 2017 install)
+* __An adequate GPU__: CARLA aims for realistic simulations, so the server needs at least a 4GB GPU. If VR is required, a Turing based GPU (for example the RTX branch of NVIDIA) is
+  highly recommended!
+* __Two TCP ports and good internet connection__: 2000 and 2001 by default. Be sure neither the firewall nor any other application is blocking these. 
 
-# 1. Compile the Python API Client
-!!! Note 
-    (Tested in June 2022) When trying to run:
-    ```commandline
+---
+
+## Required Software 
+
+### Before you continue
+
+Please make sure you have the following programs installed before continuing. If you will be installing them now, please read the instructions below carefully to prevent problems later on in the installation process.
+
+* __[CMake](https://cmake.org/download/){target="_blank"}__: A small software package to make sure the CARLA build can compile C-code. Please follow the link, scroll down and download the "Windows win64-x64 Installer" (there is no need to download the source). When installing select the option "Add CMake to the system `PATH` for all users"
[…]
