+++ modules/controllerplotter/controllerplotter_dialog.py
@@ -73,8 +73,8 @@ class ControllerPlotterDialog(ModuleDialog):
         self.plot_data_loha_stiffness_x = [-160, 160]
         self.plot_data_loha_stiffness_shifted = []
         self.plot_data_loha_stiffness_y = [0, 0]
-        self.converted_x_road_outer = [0] * 50
-        self.converted_y_road_outer = [0] * 50
+        self.converted_x_road_outer = [] * 50
+        self.converted_y_road_outer = [] * 50
 
         # change labelfont
         self.labelfont = QtGui.QFont()
@@ -118,7 +118,6 @@ class ControllerPlotterDialog(ModuleDialog):
         self.hcr_plot_handle = None
         self.auto_position_plot_handle = None
         self.road_plot_handle = None
-        self.road_plot_handle_cubic = None
         self.road_outer_plot_handle = None
         self.road_inner_plot_handle = None
         self.topview_lat_error_plot_handle = None
@@ -169,12 +168,12 @@ class ControllerPlotterDialog(ModuleDialog):
             plot_data_hcr_x = tmp.values[:, 1]
             plot_data_hcr_y = tmp.values[:, 2]
 
-            # self.hcr_plot_handle = self.module_widget.top_view_graph.plot(x=plot_data_hcr_x, y=plot_data_hcr_y, shadowPen=pg.mkPen(10, 200, 0, 100, width=18),
-            #                                                               pen=pg.mkPen(0, 102, 0, 255, width=2))
+            self.hcr_plot_handle = self.module_widget.top_view_graph.plot(x=plot_data_hcr_x, y=plot_data_hcr_y, shadowPen=pg.mkPen(10, 200, 0, 100, width=18),
+                                                                          pen=pg.mkPen(0, 102, 0, 255, width=2))
         except IOError:
             print('Could not find HCR trajectory, please hardcode a name that is in your sw controller trajectory list ')
-            # self.hcr_plot_handle = self.module_widget.top_view_graph.plot(x=[0], y=[0], shadowPen=pg.mkPen(10, 200, 0, 100, width=18),
-            #                                                               pen=pg.mkPen(0, 102, 0, 255, width=2))
+            self.hcr_plot_handle = self.module_widget.top_view_graph.plot(x=[0], y=[0], shadowPen=pg.mkPen(10, 200, 0, 100, width=18),
+                                                                          pen=pg.mkPen(0, 102, 0, 255, width=2))
         self.car_symbol = QtGui.QPainterPath()
         self.car_symbol.addRect(-0.2, -0.4, 0.4, 0.8)
 
@@ -184,11 +183,6 @@ class ControllerPlotterDialog(ModuleDialog):
                                                                                 symbolPen=pg.mkPen((0, 0, 0, 255), width=2))
         self.road_plot_handle = self.module_widget.top_view_graph.plot(x=[0], y=[0], size=2,
                                                                        pen=pg.mkPen((0, 0, 0, 200), width=1, style=QtCore.Qt.DashLine),
-                                                                       brush='g', symbol=None,
-                                                                       symbolBrush=self.brushes,
-                                                                       symbolPen=self.pens, symbolSize=5)
-        self.road_plot_handle_cubic = self.module_widget.top_view_graph.plot(x=[0], y=[0], size=2,
-                                                                       pen=pg.mkPen((0, 0, 0, 100), width=2, style=QtCore.Qt.DashLine),
                                                                        brush='g', symbol=None,
[…]
