+++ site/requirements.txt
@@ -1,0 +1,3 @@
+mkdocs >= 1.0
+pymdown-extensions
+markdown-inline-graphviz-extension>=1.1
