+++ core/hq/hq_manager.py
@@ -1,7 +1,5 @@
 """Action class for JOAN menu"""
 import os
-import subprocess
-import time
 
 from PyQt5 import QtCore
 from PyQt5.QtCore import pyqtSignal
@@ -28,9 +26,6 @@ class HQManager(QtCore.QObject):
         super(QtCore.QObject, self).__init__()
 
         self.central_state_monitor = CentralStateMonitor()
-
-        # Start Carla
-        # self.start_Carla_if_closed()
 
         # News
         self.news = News()
