### 2022-08-15 ece8ff4
Removed default_settings option. Utilize experiment manager to use default settings.
