### 2022-08-15 9ff61eb
Added checks for both vehicles checking for empty setting dicts to prevent crashing of the code (which expected non-empty dicts)
