### 2022-08-15 c08782d
Removed print() statement
