### 2022-08-15 e69109d
re-added the code that updates the modules in the central settings.
