+++ modules/carlainterface/carlainterface_process.py
@@ -11,7 +11,6 @@ from modules.joanmodules import JOANModules
 if platform.system() == 'Windows':
     import wres
 
-# from tools.carlaimporter import carla
 import carla
 
 def connect_carla(host='localhost', port=2000):
@@ -86,6 +85,7 @@ class CarlaInterfaceProcess(ModuleProcess):
 
         # Now we create our agents and directly spawn them
         for key, value in self._settings_as_object.agents.items():
+            print(value.agent_type)
             self.agent_objects[key] = AgentTypes(value.agent_type).process(self, settings=value, shared_variables=self._module_shared_variables.agents[key])
             time.sleep(0.1)  # short sleep, such that the camera in CARLA is attached properly to the first spawned car
 
