### 2022-08-17 2a2852c
Changed textboxes to spinboxes for both controllers
