### 2022-08-17 31736ea
Changed textboxes to spinboxes for both controllers
