### 2022-08-17 dc07aad
Some aesthetic changes to the code for the controllers
