### 2022-08-17 e52fa9c
Added cruise control value to shared variables to assure correct speed of npc's
