### 2022-08-17 21df277
Added cruise control value to shared variables to assure correct speed of npc's
