### 2022-08-17 e0c5f18
Update npc_vehicle.py
removed an unused print statement
