+++ modules/carlainterface/carlainterface_agentclasses/npc_vehicle.py
@@ -156,7 +156,6 @@ class NPCVehicleProcess:
             self._control.hand_brake = self.npc_controller_shared_variables.controllers[self.settings.selected_npc_controller].handbrake
             self._control.brake = self.npc_controller_shared_variables.controllers[self.settings.selected_npc_controller].brake
             self._control.throttle = self.npc_controller_shared_variables.controllers[self.settings.selected_npc_controller].throttle
-            # print(self._control.throttle)
 
             self.spawned_vehicle.apply_control(self._control)
 
