### 2022-09-01 b8a5a65
Updated tradedcontroller.py and fdcacontroller.py, should work again. Main changes in how settings are saved.
