### 2022-09-01 79e540e
Updated tradedcontroller.py and fdcacontroller.py, should work again. Main changes in how settings are saved.
