### 2022-10-05 #4250 @OlgerSiebinga
Hi All,
A quick update from our side. We have updated JOAN to support Carla 0.9.13. This new version of JOAN (v1.2.0) is currently on the develop branch waiting for this review to finish before we merge it in the main branch. This update to Carla 0.9.13 means that the installation procedure does no longer depend on visual studio 2017. Which should make installation a bit easier.

Are there any updates regarding the review?
