### 2022-10-06 #4250 @humanfactors
Hi @OlgerSiebinga

In my (somewhat limited) experience, I would say this is an abnormal delay in your paper, but indeed the complexity of the build is part of the issue. I could be wrong (@taless474 can confirm) but I changed the version of my submission for my JOSS paper — so I believe you could adjust your build while this is under review. You need to document and adjust the post details to make clear that there is a new version being submitted that addresses critiques of the review.

I'm happy to attempt to build again from the develop branch, under the assumption that a secondary reviewer is not required. My concern is spending the time to reproduce and approve — but end up still needing a second reviewer, which wouldn't be a productive immediate action to get this through.

Good luck :)
