### 2022-12-13 #4250 @humanfactors
Awesome we have another reviewer. I must state that I've lost access to my Visual Studio Installation and old laptop configured. However, as stated, I had the actual JOAN software working fine. As long as one reviewer can validate that the link between JOAN and the simulator is functional, I am personally happy to accept this.
