### 2022-12-26 #4250 @imcatta
Hi @OlgerSiebinga,
the following is the list of think I notice while performing the review. I will keep it updated as I proceed.

### Minor points
1. A link in the README is broken. I've submitted a pull request to fix it tud-hri/joan#23

### Other (non-blocking)
1. The installation procedure is very well documented. However, the user is asked to manually download the entire repository from GitHub and install the Python dependencies. Authors should consider automating the process more, such as by distributing a PIP package and/or providing the ability to download an executable (e.g., via PyInstaller). This can be considered as not blocking because installation instructions are crystal clear and thus satisfy JOSS requirements.
