### 2023-01-03 #4250 @OlgerSiebinga
Hi @imcatta,

Thanks for the fix in the pull request and the feedback in general. I've got some short replies here.

1. We deliberately ask users to fork and download the repository manually. This choice was made because JOAN is set up so that it can be used for straightforward experiments out of the box. But if you want to do anything new, you will have to make some adjustments to the code (add modules, or even add a simple scenario). So both pip and pyinstaller would simplify the installation, but would severely limit the usability.

2. Thanks for the tip, I'll look into it. Sounds very useful.

3. We have updated JOAN to work with CARLA 0.9.13. This new version of JOAN is waiting on the develop branch because the review was already pending when we implemented this.

4. I fully agree, we mostly lack the time (and funding) to implement automatic testing now. But it's high on the wish-list for new extensions.
