### 2023-01-30 #4250 @diehlpk
Hi @OlgerSiebinga please

- [ ] Generate a new release including all changes during the JOSS review and post the release number here.
- [ ] Upload the source code of the release to Zenodo or FigShare and post the DOI here. Please note that the title and the authors of the archive have to match with the paper.
