### 2023-01-30 #4250 @OlgerSiebinga
Hi @diehlpk,

I have made a new release with version number v1.1.1. The code has been uploaded to the 4TU data repository (DOI: 10.4121/21975335). This upload has to be approved by the 4TU admins, but this usually happens within a day.
